architecture: Clean-up architecture a bit
I thought a repository managing the different wallet clients was
necessary but it turns out it is better to just let the concrete wallets
be implementation of the repository. So:
- `WalletApi` is now `WalletRepository`
- `WalletApi` returns concrete model types not interfaces like
  `IntoSendPaymentResponse`. This means they also all need to convert
  the data which makes sense.
