+++ app/src/main/java/xyz/lilsus/papp/data/repository/blink/BlinkWalletRepository.kt
@@ -1,4 +1,4 @@
-package xyz.lilsus.papp.data.remote.blink
+package xyz.lilsus.papp.data.repository.blink
 
 
 import kotlinx.coroutines.Dispatchers
@@ -12,16 +12,19 @@ import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody.Companion.toRequestBody
 import xyz.lilsus.papp.common.Invoice
-import xyz.lilsus.papp.domain.model.IntoSendPaymentResult
+import xyz.lilsus.papp.domain.model.SendPaymentData
+import xyz.lilsus.papp.domain.model.SendPaymentResult
+import xyz.lilsus.papp.domain.model.config.WalletTypeEntry
 import xyz.lilsus.papp.domain.model.config.wallets.BlinkConfig
-import xyz.lilsus.papp.domain.repository.client.WalletApi
+import xyz.lilsus.papp.domain.repository.WalletRepository
+import kotlin.math.abs
 
 // FIXME: https://github.com/NicolaLS/papp-android/issues/11
 
 // Blink GraphQL API doc: https://dev.blink.sv/public-api-reference.html#mutation-lnInvoicePaymentSend
 // Don't use GraphQL client because it is not worth it.
 
-class BlinkWalletApi(config: BlinkConfig) : WalletApi {
+class BlinkWalletRepository(config: BlinkConfig) : WalletRepository {
     private val client = OkHttpClient()
     val json = Json { ignoreUnknownKeys = true }
 
@@ -60,7 +63,7 @@ class BlinkWalletApi(config: BlinkConfig) : WalletApi {
 	"""
     }
 
-    override suspend fun payBolt11Invoice(invoice: Invoice): IntoSendPaymentResult =
+    override suspend fun payBolt11Invoice(invoice: Invoice): SendPaymentResult =
         withContext(Dispatchers.IO) {
             val variables = buildJsonObject {
                 putJsonObject("input") {
@@ -97,7 +100,44 @@ class BlinkWalletApi(config: BlinkConfig) : WalletApi {
                     bodyString
                 )
 
-                return@withContext data
+                val wallet = WalletTypeEntry.BLINK
+                val payload = data.data.lnInvoicePaymentSend
+                val status =
+                    payload.status ?: return@withContext SendPaymentResult.Failure(
+                        wallet,
+                        "Missing payment status"
+                    )
+
+                return@withContext when (status) {
+                    PaymentSendResult.ALREADY_PAID -> SendPaymentResult.AlreadyPaid(wallet)
+                    PaymentSendResult.FAILURE -> {
+                        val error =
+                            "Error Sending Payment:\n ${payload.errors?.firstOrNull()?.message}"
+                        SendPaymentResult.Failure(wallet, error)
+                    }
+
+                    PaymentSendResult.PENDING -> SendPaymentResult.Pending(wallet)
+                    PaymentSendResult.SUCCESS -> {
+                        val tx =
+                            payload.transaction ?: return@withContext SendPaymentResult.Failure(
+                                wallet,
+                                "Missing transaction"
+                            )
+                        val amountPaidTotal = abs(tx.settlementAmount)
+                        val feePaid = abs(tx.settlementFee)
+                        val amountPaid = amountPaidTotal - feePaid
+                        val memo = tx.memo
+
+                        SendPaymentResult.Success(
+                            wallet = wallet,
+                            data = SendPaymentData(
+                                amountPaid = amountPaid,
+                                feePaid = feePaid,
+                                memo = memo
+                            )
+                        )
+                    }
+                }
             }
         }
 }
