+++ app/src/main/java/xyz/lilsus/papp/domain/repository/WalletRepository.kt
@@ -4,5 +4,5 @@ import xyz.lilsus.papp.common.Invoice
 import xyz.lilsus.papp.domain.model.SendPaymentResult
 
 interface WalletRepository {
-    suspend fun payBolt11Invoice(invoice: Invoice): SendPaymentResult
+    suspend fun payBolt11Invoice(bolt11Invoice: Invoice): SendPaymentResult
 }
