+++ app/src/main/java/xyz/lilsus/papp/domain/repository/WalletRepositoryFactory.kt
@@ -1,0 +1,7 @@
+package xyz.lilsus.papp.domain.repository
+
+import xyz.lilsus.papp.domain.model.config.WalletEntry
+
+interface WalletRepositoryFactory {
+    fun getClientFromConfigOrNull(walletEntry: WalletEntry): WalletRepository?
+}
