--- app/src/main/java/xyz/lilsus/papp/domain/repository/client/WalletApi.kt
@@ -1,8 +1,0 @@
-package xyz.lilsus.papp.domain.repository.client
-
-import xyz.lilsus.papp.common.Invoice
-import xyz.lilsus.papp.domain.model.IntoSendPaymentResult
-
-interface WalletApi {
-    suspend fun payBolt11Invoice(bolt11Invoice: Invoice): IntoSendPaymentResult
-}
