--- app/src/main/java/xyz/lilsus/papp/domain/repository/client/WalletClientFactory.kt
@@ -1,7 +1,0 @@
-package xyz.lilsus.papp.domain.repository.client
-
-import xyz.lilsus.papp.domain.model.config.WalletEntry
-
-interface WalletClientFactory {
-    fun getClientFromConfigOrNull(walletEntry: WalletEntry): WalletApi?
-}
