+++ app/src/test/java/xyz/lilsus/papp/data/repository/WalletRepositoryFactoryImplTest.kt
@@ -1,13 +1,13 @@
-package xyz.lilsus.papp.data.remote
+package xyz.lilsus.papp.data.repository
 
 import junit.framework.TestCase
 import org.junit.Test
-import xyz.lilsus.papp.data.remote.blink.BlinkWalletApi
+import xyz.lilsus.papp.data.repository.blink.BlinkWalletRepository
 import xyz.lilsus.papp.domain.model.config.WalletConfigEntry
 import xyz.lilsus.papp.domain.model.config.WalletEntry
 import xyz.lilsus.papp.domain.model.config.wallets.BlinkConfig
 
-class WalletClientFactoryImplTest {
+class WalletRepositoryFactoryImplTest {
     @Test
     fun walletClientFactory_returns_correct_client_instance() {
         val noWalletConfigured =
@@ -20,10 +20,10 @@ class WalletClientFactoryImplTest {
                 )
             )
 
-        val walletClientFactory = WalletClientFactoryImpl()
+        val walletClientFactory = WalletRepositoryFactoryImpl()
 
         TestCase.assertNull(walletClientFactory.getClientFromConfigOrNull(noWalletConfigured))
-        TestCase.assertTrue(walletClientFactory.getClientFromConfigOrNull(blinkWalletConfigured) is BlinkWalletApi)
+        TestCase.assertTrue(walletClientFactory.getClientFromConfigOrNull(blinkWalletConfigured) is BlinkWalletRepository)
     }
 
 }
